Separate OS impl into files, support windows and adjust paths
- I seperated platform specific functionality into files that get built
  using the build constraints of Go
- The codebase is fairly liberal with appending '/' and prepending '/'
  paths as well as some dodgy '/<name>' checks that'd break on Windows
  because the path separators are different. I've made this agnostic to
  the OS by using os.PathSeparator.

  However there are still many instances left, in some cases paths
  emitted to Windows with '/' are typically permitted by most APIs. But
  some of the logic in chifra uses Go's path libraries which is OS
  agnostic but then does string compares with a OS specific path
  seperator causing various nesting of folders bugs.
- Constructing URLs by appending the IPFS hash onto the URL was broken
  on Windows because we used 'filepath.Join'. This is similar behaviour
  as I previously mentioned where Go is agnostic and emits a '\' to the
  path instead of a '/' because we are constructing a URL which broke
  CURL requests on Windows.

  I've fixed this just by appending the strings together.
- There is some OS functionality which I've stubbed out that needs to be
  implemented like using LockFileEx, UnlockFile & co on Windows to
  replace POSIX's flock.
