+++ src/apps/chifra/pkg/file/lock_windows.go
@@ -1,0 +1,17 @@
+// Copyright 2021 The TrueBlocks Authors. All rights reserved.
+// Use of this source code is governed by a license that can
+// be found in the LICENSE file.
+
+package file
+
+import (
+	"os"
+)
+
+func Lock(file *os.File) error {
+	return nil
+}
+
+func Unlock(file *os.File) error {
+	return nil
+}
