+++ src/apps/chifra/pkg/progress/scanbar_windows.go
@@ -1,0 +1,9 @@
+// Copyright 2021 The TrueBlocks Authors. All rights reserved.
+// Use of this source code is governed by a license that can
+// be found in the LICENSE file.
+
+package progress
+
+func screenWidth() uint {
+	return 120 // default reasonably
+}
